pyproject.toml

[project]
name = "mcp-ask-llm"
version = "0.1.0"
description = "An MCP server to ask LLMs"
requires-python = ">=3.8"
dependencies = [
    "fastmcp",
    "litellm",
    "pyyaml",
    "python-dotenv",
]

[project.scripts]
start-mcp-server = "server:main"
